pyproject.toml
[tool.poetry]
name = "big-query-kusto-client"
version = "0.3.0"
description = "Wrapper around azure-kusto client to facilitate bringing BIG query results while avoiding hitting the limits"
authors = ["JJ Triff <jjtriff@gmail.com>"]
readme = "README.md"
packages = [{include = "big_query_kusto_client", from = "src"}]
license = "Apache-2.0"
homepage = "https://github.com/jjtriff/big-query-kusto-client"

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.5.2"
tenacity = "^8.2.3"
azure-kusto-data = "^4.2.0"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[project.urls]

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
